moved types file to services and deleted old api
