+++ frontend/src/components/projects/project-workers-management-dialog.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState, useCallback, useRef } from 'react'
-import { Project, Worker } from '@/api/types'
+import { Project, Worker } from '@/services/types'
 import { toast } from 'sonner'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Button } from '@/components/ui/button'
